app/requirements.txt
pip>=21.0
termcolor>=1.1.0
fastapi
uvicorn
gunicorn
httpx>=0.21.0
pytest
python-dotenv
pandas
starlette==0.27.0
sse-starlette==1.6.5
boto3==1.34.11
botocore==1.34.11